Security improvement

+++ src/Controller/GameMessageController.php
@@ -22,6 +22,9 @@ class GameMessageController extends AbstractController
     public function new(Request $request,Game $game): Response
     {
         $gameMessage = new GameMessage();
+
+        $this->denyAccessUnlessGranted('ROLE_USER', $gameMessage);
+
         $form = $this->createForm(GameMessageType::class, $gameMessage);
         $form->handleRequest($request);
 
@@ -54,6 +57,8 @@ class GameMessageController extends AbstractController
      */     
     public function edit(Request $request, GameMessage $gameMessage): Response
     {
+        $this->denyAccessUnlessGranted('MESSAGE_EDIT',$gameMessage);
+
         $form = $this->createForm(GameMessageType::class, $gameMessage);
         $form->handleRequest($request);
         
@@ -69,7 +74,6 @@ class GameMessageController extends AbstractController
         }
 
         return $this->render('game_message/edit.html.twig', [
-            'gameMessage' => $gameMessage,
             'form' => $form->createView(),
         ]);
     }
@@ -80,6 +84,8 @@ class GameMessageController extends AbstractController
      */
     public function delete(Request $request, GameMessage $gameMessage): Response
     {
+        $this->denyAccessUnlessGranted('MESSAGE_EDIT',$gameMessage);
+
         if ($this->isCsrfTokenValid('delete'.$gameMessage->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($gameMessage);
